refactor: update component structure
