+++ src/render.js
@@ -5,6 +5,7 @@ const RenderPosition = {
   AFTEREND: 'afterend',
 };
 
+
 function createElement(template) {
   const newElement = document.createElement('div');
   newElement.innerHTML = template;
@@ -12,8 +13,10 @@ function createElement(template) {
   return newElement.firstElementChild;
 }
 
+
 function render(component, container, place = RenderPosition.BEFOREEND) {
   container.insertAdjacentElement(place, component.getElement());
 }
+
 
 export { RenderPosition, createElement, render };
