+++ src/view/edit-point/components/destination-edit-view.js
@@ -1,0 +1,31 @@
+import { createElement } from '../../../render.js';
+
+
+const createDestinationEditTemplate = () => (`
+  <section class="event__section  event__section--destination">
+    <h3 class="event__section-title  event__section-title--destination">Destination</h3>
+    <p class="event__destination-description">
+    Geneva is a city in Switzerland that lies at the southern tip of expansive
+    Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.
+    </p>
+  </section>
+`);
+
+
+export default class DestinationEditView {
+  getTemplate() {
+    return createDestinationEditTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
