+++ src/view/edit-point/components/offer-list-edit-view.js
@@ -1,0 +1,79 @@
+import { createElement } from '../../../render.js';
+
+
+const createOfferListEditTemplate = () => (`
+  <section class="event__section  event__section--offers">
+    <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+
+    <div class="event__available-offers">
+      <div class="event__offer-selector">
+        <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox"
+               name="event-offer-luggage">
+        <label class="event__offer-label" for="event-offer-luggage-1">
+          <span class="event__offer-title">Add luggage</span>
+          &plus;&euro;&nbsp;
+          <span class="event__offer-price">30</span>
+        </label>
+      </div>
+
+      <div class="event__offer-selector">
+        <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox"
+               name="event-offer-comfort">
+        <label class="event__offer-label" for="event-offer-comfort-1">
+          <span class="event__offer-title">Switch to comfort class</span>
+          &plus;&euro;&nbsp;
+          <span class="event__offer-price">100</span>
+        </label>
+      </div>
+
+      <div class="event__offer-selector">
+        <input class="event__offer-checkbox  visually-hidden" id="event-offer-meal-1" type="checkbox"
+               name="event-offer-meal">
+        <label class="event__offer-label" for="event-offer-meal-1">
+          <span class="event__offer-title">Add meal</span>
+          &plus;&euro;&nbsp;
+          <span class="event__offer-price">15</span>
+        </label>
+      </div>
+
+      <div class="event__offer-selector">
+        <input class="event__offer-checkbox  visually-hidden" id="event-offer-seats-1" type="checkbox"
+               name="event-offer-seats">
+        <label class="event__offer-label" for="event-offer-seats-1">
+          <span class="event__offer-title">Choose seats</span>
+          &plus;&euro;&nbsp;
+          <span class="event__offer-price">5</span>
+        </label>
+      </div>
+
+      <div class="event__offer-selector">
+        <input class="event__offer-checkbox  visually-hidden" id="event-offer-train-1" type="checkbox"
+               name="event-offer-train">
+        <label class="event__offer-label" for="event-offer-train-1">
+          <span class="event__offer-title">Travel by train</span>
+          &plus;&euro;&nbsp;
+          <span class="event__offer-price">40</span>
+        </label>
+      </div>
+    </div>
+  </section>
+`);
+
+
+export default class OfferListEditView {
+  getTemplate() {
+    return createOfferListEditTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
