+++ src/view/edit-point/edit-point-view.js
@@ -1,0 +1,23 @@
+import { createElement } from '../../render.js';
+
+
+const createEditPointTemplate = () => '<form class="event event--edit" action="#" method="post"></form>';
+
+
+export default class EditPointView {
+  getTemplate() {
+    return createEditPointTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
