+++ src/view/filter-list/filter-list-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import { createElement } from '../../render.js';
 
 
 const createFilterListTemplate = () => (`
