+++ src/view/sort-list/sort-list-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import { createElement } from '../../render.js';
 
 
 const createSortListTemplate = () => (`
@@ -9,8 +9,7 @@ const createSortListTemplate = () => (`
     </div>
 
     <div class="trip-sort__item  trip-sort__item--event">
-      <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-             value="sort-event" disabled>
+      <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event">
       <label class="trip-sort__btn" for="sort-event">Event</label>
     </div>
 
@@ -20,14 +19,12 @@ const createSortListTemplate = () => (`
     </div>
 
     <div class="trip-sort__item  trip-sort__item--price">
-      <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-             value="sort-price" checked>
+      <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price">
       <label class="trip-sort__btn" for="sort-price">Price</label>
     </div>
 
     <div class="trip-sort__item  trip-sort__item--offer">
-      <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort"
-             value="sort-offer" disabled>
+      <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer">
       <label class="trip-sort__btn" for="sort-offer">Offers</label>
     </div>
   </form>
