+++ src/view/title-info/title-info-view.js
@@ -1,7 +1,7 @@
-import { createElement } from '../render.js';
+import { createElement } from '../../render.js';
 
 
-const createTitleTemplate = () => (`
+const createTitleInfoTemplate = () => (`
   <section class="trip-main__trip-info  trip-info">
     <div class="trip-info__main">
       <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
@@ -16,9 +16,9 @@ const createTitleTemplate = () => (`
 `);
 
 
-export default class TitleView {
+export default class TitleInfoView {
   getTemplate() {
-    return createTitleTemplate();
+    return createTitleInfoTemplate();
   }
 
   getElement() {
